# Fix enumerator allocation in TargetGroupingTransformTracker