--- a/osu.Framework/Graphics/Transforms/TargetGroupingTransformTracker.cs
+++ b/osu.Framework/Graphics/Transforms/TargetGroupingTransformTracker.cs
@@ -41,6 +41,8 @@ namespace osu.Framework.Graphics.Transforms
         /// </summary>
         private readonly Dictionary<string, int> lastAppliedTransformIndices = new Dictionary<string, int>();
 
+        private int? lastAppliedTransforIndexMin;
+
         /// <summary>
         /// All <see cref="Transform.TargetMember"/>s which are handled by this tracker.
         /// </summary>
@@ -333,7 +335,7 @@ namespace osu.Framework.Graphics.Transforms
         private int getLastAppliedIndex(string targetMember = null)
         {
             if (targetMember == null)
-                return lastAppliedTransformIndices.Values.Min();
+                return lastAppliedTransforIndexMin ?? 0;
 
             if (lastAppliedTransformIndices.TryGetValue(targetMember, out int val))
                 return val;
@@ -349,6 +351,9 @@ namespace osu.Framework.Graphics.Transforms
         private void setLastAppliedIndex(string targetMember, int index)
         {
             lastAppliedTransformIndices[targetMember] = index;
+
+            // this is safe because it's guaranteed we call this function with an equal or higher value for the targetMember.
+            lastAppliedTransforIndexMin = Math.Min(lastAppliedTransforIndexMin ?? int.MaxValue, index);
         }
 
         /// <summary>
@@ -358,6 +363,7 @@ namespace osu.Framework.Graphics.Transforms
         {
             foreach (var tracked in targetMembers)
                 lastAppliedTransformIndices[tracked] = 0;
+            lastAppliedTransforIndexMin = null;
         }
     }
 }