--- a/osu.Framework/Graphics/Transforms/TargetGroupingTransformTracker.cs
+++ b/osu.Framework/Graphics/Transforms/TargetGroupingTransformTracker.cs
@@ -41,8 +41,6 @@ namespace osu.Framework.Graphics.Transforms
         /// </summary>
         private readonly Dictionary<string, int> lastAppliedTransformIndices = new Dictionary<string, int>();
 
-        private int? lastAppliedTransformIndexMin;
-
         /// <summary>
         /// All <see cref="Transform.TargetMember"/>s which are handled by this tracker.
         /// </summary>
@@ -335,7 +333,17 @@ namespace osu.Framework.Graphics.Transforms
         private int getLastAppliedIndex(string targetMember = null)
         {
             if (targetMember == null)
-                return lastAppliedTransformIndexMin ?? 0;
+            {
+                int min = int.MaxValue;
+
+                foreach (int i in lastAppliedTransformIndices.Values)
+                {
+                    if (i < min)
+                        min = i;
+                }
+
+                return min;
+            }
 
             if (lastAppliedTransformIndices.TryGetValue(targetMember, out int val))
                 return val;
@@ -351,9 +359,6 @@ namespace osu.Framework.Graphics.Transforms
         private void setLastAppliedIndex(string targetMember, int index)
         {
             lastAppliedTransformIndices[targetMember] = index;
-
-            // this is safe because it's guaranteed we call this function with an equal or higher value for the targetMember.
-            lastAppliedTransformIndexMin = Math.Min(lastAppliedTransformIndexMin ?? int.MaxValue, index);
         }
 
         /// <summary>
@@ -363,7 +368,6 @@ namespace osu.Framework.Graphics.Transforms
         {
             foreach (var tracked in targetMembers)
                 lastAppliedTransformIndices[tracked] = 0;
-            lastAppliedTransformIndexMin = null;
         }
     }
 }