Running existing benchmark `BenchmarkManySpinningBoxes`

Before:

|   Method |     Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |---------:|---------:|---------:|-------:|------:|------:|----------:|
| RunFrame | 987.0 us | 19.31 us | 18.96 us | 7.8125 |     - |     - |  39.12 KB |

After:


|   Method |     Mean |    Error |   StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |---------:|---------:|---------:|------:|------:|------:|----------:|
| RunFrame | 995.4 us | 17.27 us | 29.80 us |     - |     - |     - |       1 B |


![image](https://user-images.githubusercontent.com/191335/95877801-76e3a380-0daf-11eb-8877-d753d5cff733.png)