@bdach 2020-10-13
Re-ran benchmarks (`BenchmarkManySpinningBoxes`) after the latest changes to double check. Still looks good.

`master`:

|   Method |     Mean |     Error |    StdDev |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |---------:|----------:|----------:|--------:|------:|------:|----------:|
| RunFrame | 1.297 ms | 0.0269 ms | 0.0746 ms | 11.7188 |     - |     - |   39.1 KB |

This PR:

|   Method |     Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |---------:|----------:|----------:|------:|------:|------:|----------:|
| RunFrame | 1.271 ms | 0.0360 ms | 0.1043 ms |     - |     - |     - |       3 B |

Not sure where the *two* extra bytes are coming from? Nevertheless a huge improvement.